Make LPC working
This will allow you to have a proper LPC device registered under OS X.
The problem is that AppleLPC.kext which is the driver has no entry for
the ICH7 device ID which is 0x27b8 in IONameMatch in Info.plist. So
LPC does not work.
This problem can be fixed by adding proper device ID in Info.plist or
by patching the DSDT like this commit. This commit fakes the device ID
to be an ICH7-M one. Its ID is 0x27b9.

Reference
http://www.efixusers.com/showthread.php?t=533